feat: Expose ComposerGraph for custom composers
