+++ src/Umbraco.Core/Composing/IComposer.cs
@@ -1,14 +1,8 @@
-using Umbraco.Cms.Core.DependencyInjection;
-
 namespace Umbraco.Cms.Core.Composing;
 
 /// <summary>
 ///     Represents a composer.
 /// </summary>
-public interface IComposer : IDiscoverable
+public interface IComposer : ICustomComposer
 {
-    /// <summary>
-    ///     Compose.
-    /// </summary>
-    void Compose(IUmbracoBuilder builder);
 }
